Debugging of the compute_fvs function (and the rest).
The commit contains an example graph that can be read with read_graph and used for testing purposes. It is a planar graph and the smalles fvs has size 3 (f.e. 2,4,7).

I couldnt find any bugs left, so for the moment I assume that the algorithm is working. Sadly the code got a little messy over yesterdays and todays debugging sessions. I decided to only comment the cout lines, if somebody else wants to use them...